"Improve unsigned conversion and update dependencies"
In this commit, we have updated the Kotlin Jps plugin and the Gradle version for handling new features and performance improvements. We also transitioned from a deprecated direct character conversion method to an advised two-step process in the UByte, UShort, ULong, and UInt files to adhere to Kotlin's best practices. The test suite and other project dependencies were also updated to their latest stable versions.

+++ src/test/kotlin/unsigned/uint.kt
@@ -25,7 +25,7 @@ class `Uint test` : StringSpec() {
             a.toFloat() shouldBe 0xffff_ffff.toFloat()
             a.toLong() shouldBe 0xffff_ffff
             a.toInt() shouldBe 0xffff_ffff.toInt()
-            a.toChar() shouldBe 0xffff_ffff.toInt().toChar()
+            a.toInt().toChar() shouldBe 0xffff_ffff.toInt().toChar()
             a.toShort() shouldBe 0xffff_ffff.s
             a.toByte() shouldBe 0xffff_ffff.b
